move functionality to the test router, give each block its own event
