+++ service_test_case.py
@@ -58,8 +58,6 @@ class NioServiceTestCase(NIOTestCase):
         self._publisher_event = Event()
         # Allow tests to publish signals to any subscriber
         self._publishers = {}
-        # Set an event when blocks process signals
-        self._processed_event = Event()
 
     @property
     def processed_signals(self):
@@ -76,7 +74,8 @@ class NioServiceTestCase(NIOTestCase):
     def publish_signals(self, topic, signals):
         self._publishers[topic].send(signals)
 
-    def notify_signals(self, block_name, signals, terminal="__default_terminal_value"):
+    def notify_signals(self, block_name, signals,
+                       terminal="__default_terminal_value"):
         self._router.notify_signals(
             self._blocks[block_name], signals, terminal)
 
@@ -104,7 +103,6 @@ class NioServiceTestCase(NIOTestCase):
         self.service_config = self.service_configs.get(self.service_name, {})
         self._setup_blocks()
         self._setup_pubsub()
-        self._setup_processed()
         # Start blocks
         if self.auto_start:
             self.start()
@@ -249,39 +247,16 @@ class NioServiceTestCase(NIOTestCase):
             block_config[property] = new_block_config[property]
         return block_config
 
-    def _processed_signals(self):
-        self._processed_event.set()
-        self._processed_event.clear()
-
-    def _call_processed(self, process_signals):
-        def process_wrapper(*args, **kwargs):
-            process_signals(*args, **kwargs)
-            self._processed_signals()
-        return process_wrapper
-
-    def _setup_processed(self):
-        """wrap every block's (including mocked blocks) process_signals
-        function with a custom one that calls _processed_signals upon exit.
-        """
-        for block_name, block in self._blocks.items():
-            block.process_signals = self._call_processed(block.process_signals)
-
     def wait_for_processed_signals(self, block_name, count=0, timeout=1):
         """ Wait the given timeout for the given block's number of processed
         signals to be equal to count.
         """
         if not count:
-            self._processed_event.wait(timeout)
+            self._blocks[block_name]._processed_event.wait(timeout)
         else:
             while count > len(self._router._processed_signals[block_name]):
-                if not self._processed_event.wait(timeout):
+                if not self._blocks[block_name]._processed_event.wait(timeout):
                     return
-
-    def wait_for_next_processed_signal(self, timeout=1):
-        """ Wait the given timeout for any block in the service to process a
-        signal.
-        """
-        self._processed_event.wait(timeout)
 
     def wait_for_published_signals(self, count=0, timeout=1):
         """Wait for the specified number of signals to be published
